+++ mullvad-daemon/src/settings.rs
@@ -16,7 +16,13 @@ use talpid_core::logging::windows::log_sink;
 
 pub fn load() -> Settings {
     match Settings::load() {
-        Ok(settings) => settings,
+        Ok(mut settings) => {
+            // Force IPv6 to be enabled on Android
+            if cfg!(target_os = "android") {
+                let _ = settings.set_enable_ipv6(true);
+            }
+            settings
+        }
         #[cfg(windows)]
         Err(SettingsError::ReadError(ref _path, ref e)) if e.kind() == ErrorKind::NotFound => {
             info!(